requirements.txt
alabaster==1.0.0
amqp==5.3.1
asgiref==3.8.1
babel==2.17.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.10
Django==5.1.5
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
docutils==0.21.2
git-filter-repo==2.47.0
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.5
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
myst-parser==4.0.1
nose==1.3.7
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
roman-numerals-py==3.0.0
six==1.17.0
snowballstemmer==2.2.0
Sphinx==8.2.0
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yt-dlp==2025.3.31
